added mmwave module

+++ src/lte/test/lte-test-pf-ff-mac-scheduler.h
@@ -33,14 +33,14 @@ using namespace ns3;
  * \ingroup lte-test
  * \ingroup tests
  *
- * \brief This system test program creates different test cases with a single eNB and 
- * several UEs, all having the same Radio Bearer specification. In each test 
- * case, the UEs see the same SINR from the eNB; different test cases are 
- * implemented obtained by using different SINR values and different numbers of 
- * UEs. The test consists on checking that the obtained throughput performance 
- * is equal among users is consistent with the definition of proportional 
- * fair scheduling
- */
+ * \brief This system test program creates different test cases with a single eNB and
+* several UEs, all having the same Radio Bearer specification. In each test
+* case, the UEs see the same SINR from the eNB; different test cases are
+* implemented obtained by using different SINR values and different numbers of
+* UEs. The test consists on checking that the obtained throughput performance
+* is equal among users is consistent with the definition of proportional
+* fair scheduling
+*/
 class LenaPfFfMacSchedulerTestCase1 : public TestCase
 {
 public:
@@ -53,7 +53,7 @@ public:
    * \param thrRefUl the UL throughput reference
    * \param errorModelEnabled if true the error model is enabled
    */
-  LenaPfFfMacSchedulerTestCase1 (uint16_t nUser, double dist, double thrRefDl, double thrRefUl, bool errorModelEnabled);
+  LenaPfFfMacSchedulerTestCase1 (uint16_t nUser, uint16_t dist, double thrRefDl, double thrRefUl, bool errorModelEnabled);
   virtual ~LenaPfFfMacSchedulerTestCase1 ();
 
 private:
@@ -63,10 +63,10 @@ private:
    * \param dist the distnace between nodes
    * \returns the name string
    */
-  static std::string BuildNameString (uint16_t nUser, double dist);
+  static std::string BuildNameString (uint16_t nUser, uint16_t dist);
   virtual void DoRun (void);
   uint16_t m_nUser; ///< number of UE nodes
-  double m_dist; ///< the distance between nodes
+  uint16_t m_dist; ///< the distance between nodes
   double m_thrRefDl; ///< the DL throughput reference
   double m_thrRefUl; ///< the UL throughput reference
   bool m_errorModelEnabled; ///< whether error model is enabled
@@ -90,7 +90,7 @@ public:
    * \param estThrPfUl the estimated UL throughput PF
    * \param errorModelEnabled if true the error model is enabled
    */
-  LenaPfFfMacSchedulerTestCase2 (std::vector<double> dist, std::vector<uint32_t> estThrPfDl, std::vector<uint32_t> estThrPfUl, bool errorModelEnabled);
+  LenaPfFfMacSchedulerTestCase2 (std::vector<uint16_t> dist, std::vector<uint32_t> estThrPfDl, std::vector<uint32_t> estThrPfUl, bool errorModelEnabled);
   virtual ~LenaPfFfMacSchedulerTestCase2 ();
 
 private:
@@ -100,11 +100,11 @@ private:
    * \param dist the distnace between nodes
    * \returns the name string
    */
-  static std::string BuildNameString (uint16_t nUser, std::vector<double> dist);
+  static std::string BuildNameString (uint16_t nUser, std::vector<uint16_t> dist);
   virtual void DoRun (void);
   uint16_t m_nUser; ///< number of UE nodes
-  std::vector<double> m_dist; ///< the distance between nodes
-  std::vector<uint32_t> m_estThrPfDl; ///< the estimated DL throughput 
+  std::vector<uint16_t> m_dist; ///< the distance between nodes
+  std::vector<uint32_t> m_estThrPfDl; ///< the estimated DL throughput
   std::vector<uint32_t> m_estThrPfUl; ///< the estimated UL throughput
   bool m_errorModelEnabled; ///< indicates whether the error model is enabled
 };
