+++ src/lte/test/lte-test-rr-ff-mac-scheduler.h
@@ -32,13 +32,13 @@ using namespace ns3;
  * \ingroup lte-test
  * \ingroup tests
  *
- * \brief This system test program creates different test cases with a single eNB and 
- * This system test program creates different test cases with a single eNB and 
- * several UEs, all having the same Radio Bearer specification. In each test 
- * case, the UEs see the same SINR from the eNB; different test cases are 
- * implemented obtained by using different SINR values and different numbers of 
- * UEs. The test consists on checking that the obtained throughput performance 
- * is equal among users and matches a reference throughput value within a given 
+ * \brief This system test program creates different test cases with a single eNB and
+ * This system test program creates different test cases with a single eNB and
+ * several UEs, all having the same Radio Bearer specification. In each test
+ * case, the UEs see the same SINR from the eNB; different test cases are
+ * implemented obtained by using different SINR values and different numbers of
+ * UEs. The test consists on checking that the obtained throughput performance
+ * is equal among users and matches a reference throughput value within a given
  * tolerance.
  */
 class LenaRrFfMacSchedulerTestCase : public TestCase
@@ -53,7 +53,7 @@ public:
    * \param thrRefUl the UL throughput reference
    * \param errorModelEnabled if true the error model is enabled
    */
-  LenaRrFfMacSchedulerTestCase (uint16_t nUser, double dist, double thrRefDl, double thrRefUl, bool errorModelEnabled);
+  LenaRrFfMacSchedulerTestCase (uint16_t nUser, uint16_t dist, double thrRefDl, double thrRefUl, bool errorModelEnabled);
   virtual ~LenaRrFfMacSchedulerTestCase ();
 
 private:
@@ -64,9 +64,9 @@ private:
    * \param dist the distance between UE nodes and eNodeB
    * \returns the name string
    */
-  static std::string BuildNameString (uint16_t nUser, double dist);
+  static std::string BuildNameString (uint16_t nUser, uint16_t dist);
   uint16_t m_nUser; ///< number of UE nodes
-  double m_dist; ///< the distance between nodes
+  uint16_t m_dist; ///< the distance between nodes
   double m_thrRefDl; ///< the DL throughput reference
   double m_thrRefUl; ///< the UL throughput reference
   bool m_errorModelEnabled; ///< indicates whether the error model is enabled
